v1.0.0

+++ SocialReport/external/ComboBox/ComboBox.m
@@ -70,9 +70,9 @@
 
 #pragma mark UITableViewDataSource
 
-- (float)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
+- (CGFloat)tableView:(UITableView *)tableView heightForRowAtIndexPath:(NSIndexPath *)indexPath
 {
-    return 28;
+    return 28.0;
 }
 
 - (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
